+++ gui/waves.h
@@ -17,8 +17,8 @@ public:
   VCDGraphicsView *top;
   int isScalar(DumpData data, std::string id);
   int isVector(DumpData data, std::string id);
-  void addText(float scenePos, float prevScenePos, QPainter *painter,
-               float lineHeight, int WAVES_GAP, long long prev,
-               long long prevFloat, DumpData dump, int idx, float y,
+  void addText(double scenePos, double prevScenePos, QPainter *painter,
+               double lineHeight, int WAVES_GAP, long long prev,
+               long long prevFloat, DumpData dump, int idx, double y,
                std::string prevString);
 };
